pyproject.toml
[tool.poetry]
name = "github-actions-openstack"
version = "0.1.0"
description = "Auto-scaling self-hosted GitHub Actions runners for OpenStack"
authors = ["Mohammed Naser <mnaser@vexxhost.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
flask = ">=2.2.5,<3.0.0"
flask-apscheduler = "^1.13.0"
github-webhook = "^1.0.4"
openstacksdk = "^1.5.0"
uwsgi = "^2.0.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
